requirements.txt
# Core dependencies for running the application
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=6.0
click>=8.0.0
scikit-learn>=1.0.0
psutil>=5.8.0

# Optional dependencies for logging, can be removed if not needed
loguru>=0.6.0
tqdm>=4.62.0
